+++ test/providers/openai/openai_advanced_test.dart
@@ -1,6 +1,5 @@
 import 'package:test/test.dart';
 import 'package:llm_dart/core/capability.dart';
-import 'package:llm_dart/models/assistant_models.dart';
 import 'package:llm_dart/models/file_models.dart';
 import 'package:llm_dart/models/moderation_models.dart';
 import 'package:llm_dart/providers/openai/openai.dart';